+++ Src/Visual Studio/Nuk/AuswertungNuk/nuk_lib/NkMemo.cpp
@@ -5,7 +5,7 @@ namespace nk
 	TMemo::TMemo(std::string Name, __int64 _id) noexcept
 		:
 		Component(Name, _id),
-		box_len(data.size())
+		box_len(static_cast<int>(data.size()))
 	{
 		NamedProperties["Text"] = &data;
 	}
@@ -20,12 +20,12 @@ namespace nk
 
 	void TMemo::draw(struct nk_context* ctx)
 	{
-		if (data.capacity() < (data.size() + 10))
+		if (data.capacity() < (data.size() + 10ull))
 		{
-			data.resize(data.capacity() + 64);
+			data.resize(data.capacity() + 64ull);
 		}
 		//nk_layout_row_static(ctx, this->height, ctx->current->bounds.w, 1); //height?
-		nk_edit_string(ctx, NK_EDIT_BOX, data.data(), &box_len, data.capacity(), nk_filter_default);
+		nk_edit_string(ctx, NK_EDIT_BOX, data.data(), &box_len, static_cast<int>(data.capacity()), nk_filter_default);
 	}
 
 }
